coding main

+++ java/arthmetic/operater.java
@@ -1,9 +1,8 @@
 
-
 public class operater {
     public static void main(String[] args) {
-        float x=2;
-         float y = 8;
+        double x=2;
+        double y=8;
         System.out.println(x+y);
         System.out.println(x-y);
         System.out.println(x*y);
